Extract shared coverage-targeting logic into interventions/utils.py (#328)
- Move parse_coverage, coverage helpers, and all related private functions
  from hiv_interventions.py to new stisim/interventions/utils.py
- Add age_sex_mask, coverage_to_number, compute_coverage_target as shared
  helpers; ART._get_n_to_treat and VMMC._get_n_to_circ are now one-line
  delegates
- Support age-only stratification (no Gender column required) for VMMC etc.
- Remove duplicate count() from base_interventions.py and
  gonorrhea_interventions.py — import from stisim.utils instead
