+++ pyproject.toml
@@ -32,6 +32,7 @@ classifiers = [
 dependencies = [
     "starsim>=3.3.3",
     "sciris>=3.2.9",
+    "numpy>=2.0.0",
     "optuna",
     "requests",
 ]
